refactor: move custom endpoints to src/resolvers directory
Move daos, address-enrichment, and delegation directories into
src/resolvers to consolidate all custom endpoint implementations.
